keepai/apps/server/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
supabase==2.0.3
requests==2.31.0
gunicorn==21.2.0
flask-socketio==5.3.2
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.1
flask-cors==4.0.0
clerk-sdk-python==0.1.0
flask-sock==0.7.0

# Ferramentas de desenvolvimento
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pytest==7.4.4
pytest-cov==4.1.0 